feat: Create PolygonPyramid class
This class is a copy of the PolygonPrism class but has the second face shape replaced with a single point leading to the creation of a pyramid or cone. This class has all the same methods as PolygonPrism but they have been adjusted to reflect the new format of the 3D shape.