feat: Make protected instance variables private
This was done in the PolygonPrism class because extending it to cover the PolygonPyramid class was not working. I think the issue was with the constructors being the same, but I never found a clear answer, so I copy and pasted my code. With more time, I would probably make a class that allows me to refactor all of my repeated code.